MYR-726: modular code post header
