+++ src/components-v2/atoms/PostHeader/avatar/post-avatar.interface.ts
@@ -1,0 +1,7 @@
+import {PostOrigin} from '../../../../interfaces/timeline';
+
+export type Props = {
+  origin: PostOrigin;
+  avatar?: string;
+  onClick: () => void;
+};
