+++ src/components-v2/atoms/PostHeader/cardTitle/cardTitle.interface.tsx
@@ -1,0 +1,4 @@
+export type CardTitleProps = {
+  text: string;
+  url: string;
+};
