+++ src/components-v2/atoms/PostHeader/cardTitle/cardTitle.style.ts
@@ -1,0 +1,14 @@
+import {Theme, createStyles, makeStyles} from '@material-ui/core/styles';
+
+export const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    root: {},
+    link: {
+      fontFamily: theme.typography.fontFamily,
+      fontSize: 20,
+      lineHeight: '18px',
+      color: theme.palette.text.primary,
+      textDecoration: 'none',
+    },
+  }),
+);
