+++ src/components-v2/atoms/PostHeader/postHeader.interface.ts
@@ -1,0 +1,12 @@
+import {Post} from '../../../interfaces/post';
+
+export type PostComponentProps = {
+  disable?: boolean;
+  post: Post;
+};
+
+export enum Platform {
+  twitter = 'twitter',
+  myriad = 'myriad',
+  reddit = 'reddit',
+}
