+++ src/components-v2/atoms/PostHeader/subHeader/post-sub-header.interface.ts
@@ -1,0 +1,7 @@
+import {User} from '../../../../interfaces/user';
+
+export type PostSubHeaderProps = {
+  platform: string;
+  date: Date;
+  importer?: User;
+};
